Refactor for modules
