+++ croncat/src/modules/mod.rs
@@ -3,6 +3,7 @@
 //!
 
 pub mod agent;
+pub mod factory;
+pub mod manager;
 pub mod polling;
 pub mod tasks;
-pub mod ws;
